requirements.txt
pytesseract
pillow
mss
opencv-python
pygetwindow
pywin32